Move util to leaderboard/expiration
