+++ leaderboard/expiration/errors.go
@@ -7,7 +7,7 @@
 // https://github.com/dayvson/go-leaderboard
 // Copyright © 2013 Maxwell Dayvson da Silva
 
-package util
+package expiration
 
 import "fmt"
