+++ leaderboard/expiration/util_suite_test.go
@@ -1,4 +1,4 @@
-package util_test
+package expiration_test
 
 import (
 	. "github.com/onsi/ginkgo"
@@ -7,7 +7,7 @@ import (
 	"testing"
 )
 
-func TestUtil(t *testing.T) {
+func TestExpiration(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Util Suite")
+	RunSpecs(t, "Expiration Suite")
 }
